Breaking several classes into separate files.
