+++ moveit_experimental/jog_arm/include/jog_arm/jog_ros_interface.h
@@ -43,65 +43,20 @@
 #define JOG_ARM_JOG_ROS_INTERFACE_H
 
 #include <Eigen/Eigenvalues>
-#include <control_msgs/JointJog.h>
+#include <jog_arm/collision_check_thread.h>
+#include <jog_arm/jog_arm_datastructures.h>
+#include <jog_arm/low_pass_filter.h>
 #include <moveit/move_group_interface/move_group_interface.h>
-#include <moveit/planning_scene_monitor/planning_scene_monitor.h>
-#include <moveit/robot_model_loader/robot_model_loader.h>
 #include <moveit/robot_state/robot_state.h>
 #include <rosparam_shortcuts/rosparam_shortcuts.h>
-#include <sensor_msgs/JointState.h>
 #include <sensor_msgs/Joy.h>
 #include <std_msgs/Bool.h>
 #include <std_msgs/Float64MultiArray.h>
 #include <tf2_geometry_msgs/tf2_geometry_msgs.h>
 #include <tf2_ros/transform_listener.h>
-#include <thread>
-#include <trajectory_msgs/JointTrajectory.h>
 
 namespace jog_arm
 {
-// Variables to share between threads, and their mutexes
-struct JogArmShared
-{
-  pthread_mutex_t shared_variables_mutex;
-
-  geometry_msgs::TwistStamped command_deltas;
-
-  control_msgs::JointJog joint_command_deltas;
-
-  sensor_msgs::JointState joints;
-
-  double collision_velocity_scale = 1;
-
-  // Indicates that an incoming Cartesian command is all zero velocities
-  bool zero_cartesian_cmd_flag = true;
-
-  // Indicates that an incoming joint angle command is all zero velocities
-  bool zero_joint_cmd_flag = true;
-
-  // Indicates that we have not received a new command in some time
-  bool command_is_stale = false;
-
-  // The new trajectory which is calculated
-  trajectory_msgs::JointTrajectory new_traj;
-
-  // Timestamp of incoming commands
-  ros::Time incoming_cmd_stamp = ros::Time(0.);
-
-  bool ok_to_publish = false;
-};
-
-// ROS params to be read. See the yaml file in /config for a description of each.
-struct JogArmParameters
-{
-  std::string move_group_name, joint_topic, cartesian_command_in_topic, command_frame, command_out_topic,
-      planning_frame, warning_topic, joint_command_in_topic, command_in_type, command_out_type;
-  double linear_scale, rotational_scale, joint_scale, lower_singularity_threshold, hard_stop_singularity_threshold,
-      lower_collision_proximity_threshold, hard_stop_collision_proximity_threshold, low_pass_filter_coeff,
-      publish_period, publish_delay, incoming_command_timeout, joint_limit_margin, collision_check_rate;
-  bool use_gazebo, check_collisions, publish_joint_positions, publish_joint_velocities, publish_joint_accelerations;
-};
-
 /**
  * Class JogROSInterface - Instantiated in main(). Handles ROS subs & pubs and creates the worker threads.
  */
@@ -127,33 +82,12 @@ private:
                                         const robot_model_loader::RobotModelLoaderPtr& model_loader_ptr);
 
   // Variables to share between threads
-  struct JogArmShared shared_variables_;
+  JogArmShared shared_variables_;
 
   robot_model_loader::RobotModelLoaderPtr model_loader_ptr_;
 
   // Store the parameters that were read from ROS server
-  struct JogArmParameters ros_parameters_;
-};
-
-/**
- * Class LowPassFilter - Filter the joint velocities to avoid jerky motion.
- * This is a second-order Butterworth low-pass filter.
- * See https://ccrma.stanford.edu/~jos/filters/Example_Second_Order_Butterworth_Lowpass.html
- */
-class LowPassFilter
-{
-public:
-  explicit LowPassFilter(double low_pass_filter_coeff);
-  double filter(double new_measurement_);
-  void reset(double data);
-
-private:
-  double previous_measurements_[3] = { 0., 0., 0. };
-  double previous_filtered_measurements_[2] = { 0., 0. };
-  // Larger filter_coeff-> more smoothing of jog commands, but more lag.
-  // Rough plot, with cutoff frequency on the y-axis:
-  // https://www.wolframalpha.com/input/?i=plot+arccot(c)
-  double filter_coeff_ = 10.;
+  JogArmParameters ros_parameters_;
 };
 
 /**
@@ -224,8 +158,8 @@ protected:
   tf2_ros::Buffer tf_buffer_;
   tf2_ros::TransformListener tf_listener_;
 
-  std::vector<jog_arm::LowPassFilter> velocity_filters_;
-  std::vector<jog_arm::LowPassFilter> position_filters_;
+  std::vector<LowPassFilter> velocity_filters_;
+  std::vector<LowPassFilter> position_filters_;
 
   ros::Publisher warning_pub_;
 
@@ -234,15 +168,6 @@ protected:
   const int gazebo_redundant_message_count_ = 30;
 };
 
-class collisionCheckThread
-{
-public:
-  collisionCheckThread(const JogArmParameters parameters, JogArmShared& shared_variables,
-                       const robot_model_loader::RobotModelLoaderPtr model_loader_ptr);
-};
-
 }  // namespace jog_arm
-
-static const std::string LOGNAME = "jog_arm_server";
 
 #endif  // JOG_ARM_JOG_ROS_INTERFACE_H
